refactor(User): extracted `GroupGrain` logic to  `ConnectionGroupGrain`. `UserGrain` and `GroupGrain`  inherits from `ConnectionGroupGrain`.
