+++ src/SignalR.Orleans/Groups/IGroupGrain.cs
@@ -1,13 +1,8 @@
-using System;
-using System.Threading.Tasks;
-using Orleans;
+using SignalR.Orleans.Core;
 
 namespace SignalR.Orleans.Groups
 {
-    public interface IGroupGrain : IGrainWithStringKey
+    public interface IGroupGrain : IConnectionGroupGrain
     {
-        Task AddMember(string connectionId);
-        Task RemoveMember(string connectionId);
-        Task SendMessage(object message);
     }
 }
