+++ src/SignalR.Orleans/Users/IUserGrain.cs
@@ -1,0 +1,8 @@
+﻿using SignalR.Orleans.Core;
+
+namespace SignalR.Orleans.Users
+{
+    public interface IUserGrain : IConnectionGroupGrain
+    {
+    }
+}
